requirements.txt
chainlit
load_dotenv
Optional
uvicorn